iris/requirements.txt
numpy==1.24.1
pandas==1.5.3
scikit-learn==1.2.1
python-dotenv==0.21.0
configparser==5.3.0
torch==1.13.1
pytest==7.2.1
xgboost==1.7.4